requirements.txt
numpy==1.25.2
pandas==2.1.0
matplotlib==3.8.0
dill==0.3.7
scikit-learn==1.3.0
imbalanced-learn==0.11.0
xgboost==2.0.0
mrmr_selection==0.2.8
PyYAML==6.0.1